pyproject.toml
[project]
name = "sci-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "marker-pdf>=1.7.4",
    "requests>=2.31.0",
    "streamlit>=1.45.1",
    "watchdog>=6.0.0",
]
